Split authentication concern from ApplicationController
